ml-backend/requirements.txt
# Traffic Sign Detection — ML Backend Dependencies
# Install: pip install -r requirements.txt

fastapi==0.115.*
uvicorn[standard]==0.34.*
python-multipart==0.0.*

ultralytics>=8.0.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
Pillow>=10.0.0
